+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Setup [Alembic](https://alembic.sqlalchemy.org/en/latest/) for migrations
 * Install [structlog](https://www.structlog.org/en/stable/)
 * Add a list of allowed [CORS origins](https://fastapi.tiangolo.com/tutorial/cors/)
+* Configure application for versioning
 
 ## [0.1.0](https://github.com/julleks/aviauth-api/commits/0.1.0) - 2021-12-08